[INIT] Initial scripts
This is my first run at scraping technologies from the Angelist.co
job board.  They discontinued their API as far as I can tell, so I
resorted to scraping. It is by no means meant to be pure or completely
accurate, but I think the overall counts are indicative of where our
industry is investing and spending time.